ControlioServer
This is the server that runs on the host machine and listens to commands
issued from the Android app Controlio. The commands it caters to are
specified in dictionary.properties file in /bin/in/control directory.
These commands can be issued via speech input from Controlio app or by
buttons provided on the app. Most of these commands are for navigation
on the desktop.